Update bootstrap.css
I think that .img-fill will solve the problem with images to be the same width as the container , because I have seen a lot of people trying to solve but they can't or maybe I have not found solution in google.
But I tried to solve by myself and in the end I figured out that margin-left and margin-right is the problem. 

Good Luck.